+++ bpmn_python/graph/classes/activities/__init__.py
@@ -2,4 +2,4 @@
 """
 Package init file
 """
-__all__ = ["activity_type", "subprocess_type", "task_type"]
+__all__ = ["activity.py", "subprocess.py", "task.py"]
