+++ bpmn_python/graph/classes/activities/activity.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 from pydantic import Field
 
-from bpmn_python.graph.classes.flow_node_type import FlowNode
+from bpmn_python.graph.classes.flow_node import FlowNode
 
 
 class Activity(FlowNode):
