+++ bpmn_python/graph/classes/activities/task.py
@@ -2,7 +2,7 @@
 """
 Class used for representing tTask of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.activities.activity_type import Activity
+from bpmn_python.graph.classes.activities.activity import Activity
 
 
 class Task(Activity):
