+++ bpmn_python/graph/classes/condition_expression.py
