+++ bpmn_python/graph/classes/events/event.py
@@ -2,7 +2,7 @@
 """
 Class used for representing tEvent of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.flow_node_type import FlowNode
+from bpmn_python.graph.classes.flow_node import FlowNode
 
 
 class Event(FlowNode):
