+++ bpmn_python/graph/classes/events/intermediate_catch_event.py
@@ -2,7 +2,7 @@
 """
 Class used for representing tIntermediateCatchEvent of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.events.catch_event_type import CatchEvent
+from bpmn_python.graph.classes.events.catch_event import CatchEvent
 
 
 class IntermediateCatchEvent(CatchEvent):
