+++ bpmn_python/graph/classes/events/intermediate_throw_event.py
@@ -2,7 +2,7 @@
 """
 Class used for representing tIntermediateThrowEvent of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.events.throw_event_type import ThrowEvent
+from bpmn_python.graph.classes.events.throw_event import ThrowEvent
 
 
 class IntermediateThrowEvent(ThrowEvent):
