+++ bpmn_python/graph/classes/events/start_event.py
@@ -2,7 +2,7 @@
 """
 Class used for representing tStartEvent of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.events.catch_event_type import CatchEvent
+from bpmn_python.graph.classes.events.catch_event import CatchEvent
 
 
 class StartEvent(CatchEvent):
