+++ bpmn_python/graph/classes/gateways/__init__.py
@@ -2,4 +2,4 @@
 """
 Package init file
 """
-__all__ = ["exclusive_gateway_type", "gateway_type", "inclusive_gateway_type", "parallel_gateway_type"]
+__all__ = ["exclusive_gateway.py", "gateway.py", "inclusive_gateway.py", "parallel_gateway.py"]
