+++ bpmn_python/graph/classes/gateways/exclusive_gateway.py
@@ -2,7 +2,7 @@
 """
 Class used for representing tInclusiveGateway of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.gateways.gateway_type import Gateway
+from bpmn_python.graph.classes.gateways.gateway import Gateway
 
 
 class ExclusiveGateway(Gateway):
