+++ bpmn_python/graph/classes/lane_set.py
@@ -2,11 +2,13 @@
 """
 Class used for representing tSetLane of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.base_element_type import BaseElement
 from typing import Optional, List
-from pydantic import Field, field_validator
 
-from bpmn_python.graph.classes.lane_type import Lane
+from pydantic import Field
+
+from bpmn_python.graph.classes.base_element import BaseElement
+from bpmn_python.graph.classes.lane import Lane
+
 
 class LaneSet(BaseElement):
     """
