+++ bpmn_python/graph/classes/root_element/__init__.py
@@ -2,4 +2,4 @@
 """
 Package init file
 """
-__all__ = ["callable_element_type", "event_definition_type", "process_type", "root_element_type"]
+__all__ = ["callable_element.py", "event_definition.py", "process.py", "root_element.py"]
