+++ bpmn_python/graph/classes/root_element/event_definition.py
@@ -2,7 +2,7 @@
 """
 Class used for representing tEventDefinition of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.root_element.root_element_type import RootElement
+from bpmn_python.graph.classes.root_element.root_element import RootElement
 
 
 class EventDefinition(RootElement):
