+++ bpmn_python/graph/classes/root_element/root_element.py
@@ -2,7 +2,7 @@
 """
 Class used for representing tRootElement of BPMN 2.0 graph
 """
-from bpmn_python.graph.classes.base_element_type import BaseElement
+from bpmn_python.graph.classes.base_element import BaseElement
 
 
 class RootElement(BaseElement):
